+++ DX12Engine/PageProvider.cpp
@@ -51,7 +51,7 @@ void PageProvider::addPageLoadRequestHelper(PageLoadRequest& pageRequest, Virtua
 	PageProvider& pageProvider = *pageRequest.pageProvider;
 	size_t pageRequestIndex = &pageRequest - pageProvider.pageLoadRequests;
 	ScopedFile file(resourceInfo.fileName, ScopedFile::accessRight::genericRead, ScopedFile::shareMode::writeMode, ScopedFile::creationMode::openExisting, nullptr);
-	unsigned int mipLevel = pageRequest.allocationInfo.textureLocation.mipLevel();
+	unsigned int mipLevel = (unsigned int)pageRequest.allocationInfo.textureLocation.mipLevel();
 	signed long long filePos = 0;
 	auto width = resourceInfo.widthInTexels;
 	auto height = resourceInfo.heightInTexels;
